Move plunker templates location and example data name
