+++ src/app/shared/plunker/example-data.ts
@@ -1,8 +1,9 @@
 /**
- * Plunker data
+ * Example data
  *   with information about Component name, selector, files used in example, and path to examples
+ *   TODO(tinayuangao): find a better place for this file
  */
-export class PlunkerData {
+export class ExampleData {
   // TODO: figure out how do we get these variables.
   description: string = 'Some description for material';
   // TODO: use real example and delete the example/ folder.
