+++ src/assets/template/main.ts
